Sandbox: Perform JDBC health checks on a timer. (#3773)
* sandbox: HikariJdbcConnectionProvider.start(), rather than construction.

It's too easy to construct things; I'd like it to be obvious that we're
starting something (in this case a connection pool).

* sandbox: In SqlLedgerSpec, stop ledgers before shutting down PostgreSQL.

* sandbox: Perform JDBC health checks on a timer.

The problem with hooking into existing requests is that if the Sandbox
is running in a load balancer and reports itself as unhealthy, it might
be taken out of the load balancer. Without any requests going through,
it'll never realize it's healthy again, and so will remain thinking it's
unhealthy forever.