db: loosen transaction persistence check (#3970)
Currently there is a constraint on the postgres `ledger_entries` table
ensuring that submitter, application_id and command_id are set on
transactions being persisted. These fields are coming from the submitter
info that will be empty when the transaction accepted event is received
on a participant which is not hosting the submitter.

This change updates the check to ensure that all of these fields are
set OR none of them are.

The same change has been made to the H2 migrations with the addition of
correcting the `is not null` check (was previously using `!= null`).

For postgres, as the check is just loosening the existing constraint of all fields
requiring a value I think there is no risk that the new constraint will fail
to apply to existing data. For H2 as the original constraint was
incorrect there is a possibility that existing data may be incorrect -
however as the same component is responsible for writing to H2 and postgres
I will assume the data was correctly persisted regardless of the constraint.
Plus as H2 is typically only used for ephemeral tests, there is
likely no existing data to migrate.

CHANGELOG_BEGIN

- [Sandbox] Loosens database schema to allow persistence of transaction
  ledger entries where no submitter info is present (typically when the
  submitter is hosted by another participant node).

CHANGELOG_END