Attempt to fix ghci test on MacOS (#3986)
On CI GHCi occasionally seems to get stuck on MacOS. I have no clue why that is
and I am unable to reproduce this locally. However, based on some
testing on CI this change seems to fix this and it simplifies things. The
ghci script was originally used when we still had the fat repl target
but now that we only load damlc, there is no need for this anymore.

CHANGELOG_BEGIN
CHANGELOG_END