Make successful LookupByKey a Fetch
To properly compute the informees of a lookup by key, we require the
signatories of the looked up contract. Instead of extending LookupByKey
to include the signatories, we rather emit a Fetch node instead of a
LookupByKey node on successful lookups. This also matches how this is
described in the ledger model.

To simplify the implementation the passing of token to the update
builtin functions was dropped as it is unnecessary.

The "SResultNeedKey" is now "SResultNeedContractByKey" which returns
not only the contract id, but the contract itself. Likewise for
engine's "ResultNeedKey".