temporary Windows machins for Andreas
The recent changes to the way in which we build npm packages with Bazel
have caused a lot of issues on Windows. To debug those, Andreas has
requested a temporary machine.

This is pretty much an exact replica of #3294 (a87749113), with the same
plan:

1. I run terraform apply on this PR is merged.
2. I manually, through the GCP web console, set a dummy password for that
  machine's RDP connection and transmit that to @aherrmann-da through
  Slack.
3. @aherrmann-da debugs the issue.
4. I create a PR to roll back this one, then apply it once it's merged.

Note: I have verified that master applies cleanly prior to opening this
PR.

CHANGELOG_BEGIN
CHANGELOG_END