ci cron: tell Slack on failure
Currently, when cron jobs fail, this is completely silent. After #4178,
in which we realized the VSCode extension had not been updated for about
a month, I believe it may be a good idea to get error messages sent to
Slack when a cron job fails.

We'll need to monitor for verbosity, but hopefully they will work most
of the time.

CHANGELOG_BEGIN
CHANGELOG_END