sandbox: Make sure we exit cleanly if the API server fails to start.
Previously we were only cleaning up and exiting if loading the DAML
packages failed. Now we also handle API server errors.

An example of where this might happen is if the ledger ID specified is
different to the existing ledger ID, causing an error.

One day we'll kill the ResetService and all of this will go away. One
day…

CHANGELOG_BEGIN
- [Sandbox] Fix an error that stops the server from exiting cleanly if
  API server initialization fails.
CHANGELOG_END